homeAutomation/requirements.txt
paho-mqtt==1.6.1
psycopg2-binary==2.9.7
python-json-logger==2.0.7

